+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 - Updated ChickenTeachHandler
 - Updated LiquidEggFluidWrapper to FluidEggWrapper
 - Updated Entity registration
+- MenuUtils is not able to do offsets
 
 ### Deprecated